+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Pending
 
+- [\#76](https://github.com/arkworks-rs/curves/pull/76) twisted Edwards parameters for bls12-377
+- Fixed curve benches
+
 ### Breaking changes
 
 ### Features